Decoupleing Word table and Game
Removed the association between word and game.
Word use to have many games not any more.

Since the word list is editable by users words were deleted that were assosseated with a game.

This caused a problem when going back to old games.

Now when a game is created in the DB is saves the word not the id into the game table.

Removed unnessasary columns from the games table.